Numeric cleanups

+++ src/net/coagulate/JSLBot/Handlers/Agent.java
@@ -122,8 +122,8 @@ public class Agent extends Handler {
 	@CmdHelp(description="Returns some basic information about the logged in agent")
 	public String statusCommand(final CommandEvent command) {
 		return "Agent is "+firstname+" "+lastname+"\n"+"("+grouptitle+" of "+groupname+")\n"+"Pos: "+bot.getPos()+"\n"+"Looking: "+bot.getLookAt()+"\n"+"Region: "+bot.getRegionName()+"\n"+"Bytes IN: "+bot.bytesin
-				.get()+"    OUT: "+bot.bytesout.get()+"\n"+"BPS IN: "+(int) (((float) bot.bytesin.get())/((float) bot.getSecondsSinceStartup()))+"    OUT: "+(int) (((float) bot.bytesout
-				.get())/((float) bot.getSecondsSinceStartup()));
+				.get()+"    OUT: "+bot.bytesout.get()+"\n"+"BPS IN: "+(int) (((float) bot.bytesin.get())/ bot.getSecondsSinceStartup())+"    OUT: "+(int) (((float) bot.bytesout
+				.get())/ bot.getSecondsSinceStartup());
 	}
 
 	@Nonnull
